psd2-validator/psd2-certificate-validator/src/main/java/de/adorsys/psd2/validator/certificate/util/URLDecodingUtil.java
/*
 * Copyright 2018-2022 adorsys GmbH & Co KG
 *
 * This program is free software: you can redistribute it and/or modify it
 * under the terms of the GNU Affero General Public License as published
 * by the Free Software Foundation, either version 3 of the License, or (at
 * your option) any later version. This program is distributed in the hope that
 * it will be useful, but WITHOUT ANY WARRANTY; without even the implied
 * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 * See the GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program. If not, see https://www.gnu.org/licenses/.
 *
 * This project is also available under a separate commercial license. You can
 * contact us at psd2@adorsys.com.
 */

package de.adorsys.psd2.validator.certificate.util;

import lombok.extern.slf4j.Slf4j;
import org.apache.commons.codec.DecoderException;
import org.apache.commons.codec.net.URLCodec;

@Slf4j
public class URLDecodingUtil {
    private URLDecodingUtil() {}

    public static byte[] decode(byte[] encodedCert) {
        try {
            return URLCodec.decodeUrl(encodedCert);
        } catch (DecoderException e) {
            log.debug("Error URL-decoding the data");
        }
        return new byte[0];
    }
}
